Smarter AI
Ai now weighs its options, and choses the optimal one.

Interactions now have types, which the AI uses to differentiate the effects. There is also a multiplier array, so that if we include any upgrades, we simply need to change the array values.

The evaluation could use some more thought.